Final commit, the final code for "The Weather Station"
Guys, I actually loved this hackathon so much. This was an amazing experience, all the things we did were totally worth it and iIjust loved it. I'm sad that it is already ending and I hope I can participate somewhere like this again. written with <3 by Raúl, Gabriel and Lourenço